requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
requests==2.31.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
zeep>=4.2.1
spyne>=2.14.0
lxml>=5.0.0
